Begin a ingredient list
Keep a list of known ingredients with singular and plural names,
and translations and imgages, and use these to normalize the ingredients
that are found in recipes. And populate the ingredients page from
the ingredients found in recipes.

Ingredient images can be put in $DATADIR/recipes/images, under the
normalized name.