Introduce a segmented ingredients list
This makes it possible to have separate ingredient lists
for, say, the dough and the frosting of a cake. There are
multiple small visual fixups to the edit page here too,
following a new mockup by Jakub.

The text format for the ingredients list now looks like this:

amount unit ingredient\tsegment\n