Change the way the back button works
Make the search page an 'implicit' main view, and make the back
button more or less always go back to main view. We make an exception
for the chef and category pages, which we do add to the back stack.
Also change the flow after saving a recipe. We go to the details
page for the recipe now.

https://bugzilla.gnome.org/show_bug.cgi?id=776625