Change strategy for tile backgrounds
Instead of creating a custom per-widget style provider
that needs to load the image again, generate just a single
style provider that provides the style for all recipes
and another one for all the chefs. This avoids lots of
image reloading and speeds up the list pages quite a bit.