wip: work towards better plural handling for ingredients
The goal is to get things like:
1     Egg
2 1/2 Eggs
3     Eggs
We only want to do this for ingredients that have no units, and
we want to use ngettext for this. Before we can do this we need
to clean up the distinction between IDs and display labels for
ingredients. We need to keep the ingredient list with untranslated
IDs and only translate them for presentation.