Don't show dairy warning for vegan recipes
Vegan recipes by definition can't contain dairy so we shouldn't show a
warning for it even if the milk-free flag isn't set. Another solution
would be to set the milk-free flag for all vegan recipes, but the code
seems to make the assumption that each recipe has only one diet flag
set.

https://bugzilla.gnome.org/show_bug.cgi?id=777621