cuisine-page: Add missing return sink for signal
A signal with a return type other than G_TYPE_NONE expects “parameters
to be passed to the signal, followed by a location for the return value”
but the latter was not being supplied here. This can cause a segfault
because the valist doesn’t contain the number of arguments GLib expects.

https://bugzilla.gnome.org/show_bug.cgi?id=777871