Don't show a dialog when one is already shown
The dialog windows can pile up each time chef information
was clicked in applicaton menu.

So, instead of always creating a new dialog, check if there is already one,
and create only if there is none.

https://bugzilla.gnome.org/show_bug.cgi?id=778340

We also take this opportunity to fix a few details in
the chef dialog: Always emit the done signal when the
dialog is dismissed, and leave the destruction of the
dialog to the outside code that created the dialog in
the first place.