Do less work when recipes change
We were just rebuilding all the pages, all the time, regardless
whether they are currently shown or not. This was very noticable
when using the serves spin on the shopping page - it took a
long time for the page to refresh, several hundred milliseconds.
With this change, we update in ~20.