Fix up mail prototype
We do async calls here, so we don't really have the opportunity
to return an error. Instead of going full-blown async, just
swallow the possible errors here, and simplify the callers.

At the same time, take a window argument and pass its handle
to the portal, so we have a chance to get window stacking right.