meson: Fix generation of various sources
Enumtypes are generated from gr-diet.h, not gr-recipe.h

cuisine.css is generated at configure-time from cuisine.css.in

segments.inc, ingredients.inc, no-ingredients.inc are generated using
`sed` in Autotools, but this is not cross-platform-compatible, so use
a Python script for it. This is also simpler to understand and bypasses
potential string quoting/escaping issues.

compile_resources needs a unique `c_name:` or there will be a symbol
redefinition error.

The target name of gdbus_codegen is used to generate the .h and .c
files, so use the same name as Autotools and add it to the list of
sources.