meson: Add wrap file to auto-setup libgd subproject
This is the preferred way to auto-setup subprojects. If you use a git
submodule, Meson will not know that it should initialize it before
configuration.

With this wrap file, the specified revision from the specified git URL
(or tarball, etc) will be cloned (or extracted) into the specified
directory. For more examples, see:

https://github.com/mesonbuild/meson/wiki/Wrap-dependency-system-manual