Redo our custom css handling
Instead of creating two CSS provider objects, just import
the recipes.css from the cuisine.css, since we are rewriting
cuisine.css at runtime anyway.

At the same time, introduce separate recipes-light.css and
recipes-dark.css files, so we can do color tweaks for the
dark theme in the future.

And reload the CSS when the gtk-application-prefer-dark-theme
setting changes. This is not really a common occurrence, but
it is easy to do and it helps for debugging the CSS with the
GTK+ inspector.