about dialog: Don't hardcode the recipe authors
We can get this information from the recipe database.
Doing so is a small step towards decoupling the recipe
and chef data from the application, so they can eventually
be updated independently.

Since the list returned from the database is unordered,
we now sort the list of recipe contributors alphabetically.