Rework GrImage with new loading APIs
Add an async API to GrImage. The implementation uses libsoup
to possibly load images from the network (the location where
to load images from still has to be determined).

The implementation is inspired by gnome-software's code for loading
screenshots, but is generally simpler. We do support loading a small
thumbnail first, and displaying it blurred, while the main image is
downloading.

Downloaded images are stored in XDG_CACHE_DIR/gnome-recipes.