gsettings: Make schema id == app id
It seems an open question currently if gsettings schemas
should be exported from flatpak apps or not. If we end
up deciding that they should, then having the schema ID
already identical to the app ID will make things easy,
so lets do that while we still have the chance.