Redo the event handling for the image page
Instead of forwarding events from GrWindow, grab focus to the
event box and handle events on it. This makes keynav in the
fullscreen image view work again.

This uncovered a problem with the focus grabbing we now do in
gr_edit_page_clear - it may end up grabbing the focus to an
unrealized entry, which leads to critical warnings. So, redo
this as well, by doing the focus grabbing separately, and
only do it when we are actually showing the edit page.

https://bugzilla.gnome.org/show_bug.cgi?id=779958