wip: load recipe data from the net
This is a first implementation of loading recipes.db,
chefs.db, picks.db, and recipe translations from the net.

We download a data.tar.gz and unpack it, to keep things simple.

Still to be done: reloading the newer data without restarting
the application, and checking for updates every now and then.