Avoid downloading the same data over and over
We are downloading the data as an archive, but we check the
timestamp of an extracted file. So we have to tell tar not
to extract the timestamps of files, so that they get the
timestamp from when the archive was extracted. Otherwise,
the newly extracted file will appear old, causing us to
download the same data again, next time.