Fix insertion of unit completions
The code here was just not using GtkEntryCompletion
properly. It always inserts the value that it finds
in the model, in the column indicated by text-column.
so, instead of playing games with match_func and format_func,
use a model that has the right text in it. We still try
to be clever here, and use a filter model to create
the right strings on the fly. Maybe that is not worth
doing, but it was fun.