details: Don't reload the page unnecessarily
When the notes are edited, we are saving them in the background,
causing us to reload the entire page for almost every typed
characters. This is not only excessive unnecessary work, but
also causes the text cursor in the notes popover to jump to
the end of the buffer, disrupting typing.

Fix this by avoiding the reload when we know it is unnecessary.