details: Don't preserve serves value
As far as I can see, this was always erroneous. The only way
the recipe can change in the store is if we go to the edit page
and come back. In this case, preserving the serves value can
lead to confusing behavior, where we see ingredients amounts
change unexpectedly. So just don't do it.