Complete the vulgar fraction support
We want to display all the fractions with small denominators (up
to 10) as fractions. So far, we only did so for the fractions which
are represented in Unicode. The right way to go about this would
be to use the OpenType fraction feature, but it is unfortunately
not supported well enough in our fonts. Therefore, synthesize the
remaining fractions using superscripts and subscripts, which works
with most fonts.