Redo rational approximation again
Instead of the elegant code that uses continued fractions,
use a stupid table of candidates. This produces slightly
nicer results, since it allows us to restrict the results
to 'nice' denominators, and not just denominators of a
certain size. This gets rid of 'weird' fractions like
1/7 or 1/9 or 1/13.

Update the expected results for the number tests to
reflect the approximations generated by this code.